pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "J L Gatz",email = "elan.rgb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "fast_zero", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
pytest-asyncio = "^0.26.0"

[tool.taskipy.tasks]
run = 'fastapi dev src/fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src/fast_zero -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']
exclude = [".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".history", ".mypy_cache", ".nox", ".pants.d", ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv", "__pypackages__", "_build", "buck-out", "dist", "node_modules", "venv"]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
